Finally-Audio-Server-main/Finally-Audio-Server-main/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.1
Jinja2==3.1.3

starlette==0.27.0

sqlalchemy==2.0.20

python-jose==3.3.0
python-multipart==0.0.5